+++ retrofit-response-type-keeper/src/test/resources/kotlin-suspend/Service.kt
@@ -1,0 +1,11 @@
+package test
+
+import retrofit2.*
+import retrofit2.http.*
+
+internal class Body
+
+internal interface Service {
+  @GET("/")
+  suspend fun get(c: Body): Any
+}